File organization improvement
